pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chrome-devtools-mcp-fork"
version = "2.0.0"
description = "Chrome DevTools Protocol integration for Model Context Protocol - Clean Architecture"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "withLinda13"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "websocket-client>=1.0.0",
    "requests>=2.25.0"
]

[project.urls]
Homepage = "https://github.com/withLinda/chrome-devtools-mcp-fork"
"Bug Reports" = "https://github.com/withLinda/chrome-devtools-mcp-fork/issues"
"Source" = "https://github.com/withLinda/chrome-devtools-mcp-fork"

[project.scripts]
chrome-devtools-mcp-fork = "chrome_devtools_mcp_fork.main:main"

[tool.setuptools.packages.find]
include = ["chrome_devtools_mcp_fork*"]
exclude = ["tests*"]